Tahmid Chowdhury With Personal Site Alpha
This is a demo personal website created using only HTML and CSS. I have utilized front-end frameworks along with Flexbox and Grid for layout. I refer to this as the alpha version because it is not fully responsive; while it works well on laptops and tablets, I have not yet achieved full responsiveness on mobile devices and smartwatches.

I recently re-entered the web development field after a long break, which has led me to forget many things—something that happens to many people. However, I am committed to improving myself every day and striving to surpass my previous achievements, InshaAllah. My only competitor is myself, and my greatest obstacle is also myself. I still experience some fear when it comes to working in web development, and I need to overcome this fear.